I sdded JSON files for grid levels 75-83. Also, added an extra page to UIScrollview, the extra page will show more levels coming soon sign, animating the sign around its anchor point using CMMotionmanager and UIDynamics to give it natural feel(doesn't work yet),  Also, I've been working on the WelcomeVC intro animations.

+++ HexitSpriteKit/HDHexagonControl.m
@@ -41,7 +41,7 @@ static const CGFloat kPageSpacing        = 34.0f;
 
 - (void)drawRect:(CGRect)rect
 {
-    const CGFloat kStartOriginX = ceilf(CGRectGetMidX(self.bounds) - ((self.numberOfPages - 1)/2) * kPageSpacing);
+    const CGFloat kStartOriginX = ceilf(CGRectGetMidX(self.bounds) - (((CGFloat)self.numberOfPages - 1)/2) * kPageSpacing);
     for (NSInteger page = 0; page < self.numberOfPages; page++) {
         
         CGPoint point = CGPointMake(kStartOriginX + (page * kPageSpacing), CGRectGetMidY(self.bounds));
